[Refactor] Refactor load workflow code into DataWorkflows
