+++ be/module_boundary_manifest.json
@@ -491,6 +491,7 @@
       "owned_roots": ["be/src/agent"],
       "allowed_include_prefixes": [
         "agent/",
+        "data_workflows/",
         "cache/",
         "compute_env/workgroup/",
         "exec/pipeline/query_context.h",
@@ -516,6 +517,7 @@
       ],
       "allowed_target_deps": [
         "Runtime",
+        "DataWorkflows",
         "Storage",
         "Cache",
         "ComputeEnv",
@@ -531,6 +533,90 @@
         "StarRocksGen"
       ],
       "remediation": "Keep AgentServer as FE-agent task and heartbeat orchestration below Service; move service/bootstrap integration upward."
+    },
+    {
+      "id": "dataworkflows",
+      "doc_label": "DataWorkflows",
+      "summary": "Data-plane workflow orchestration above concrete Storage and Exec for load, delete, schema-change, and tablet-write workflows without service/bootstrap ownership.",
+      "owned_targets": ["DataWorkflows"],
+      "owned_roots": ["be/src/data_workflows"],
+      "allowed_include_prefixes": [
+        "data_workflows/",
+        "storage/",
+        "exec/",
+        "runtime/",
+        "compute_env/",
+        "serde/",
+        "platform/",
+        "fs/",
+        "io/",
+        "column/",
+        "types/",
+        "common/",
+        "base/",
+        "gutil/",
+        "gen_cpp/"
+      ],
+      "forbidden_include_prefixes": [
+        "agent/",
+        "service/",
+        "connector/",
+        "http/",
+        "script/",
+        "tools/",
+        "udf/"
+      ],
+      "allowed_target_deps": [
+        "Storage",
+        "Rowset",
+        "Exec",
+        "Runtime",
+        "StorageBase",
+        "StoragePrimitive",
+        "ExecRuntime",
+        "ExecPrimitive",
+        "ComputeEnv",
+        "Serde",
+        "RuntimeEnv",
+        "RuntimeCore",
+        "Platform",
+        "FSCore",
+        "IO",
+        "ChunkCore",
+        "ColumnCore",
+        "Types",
+        "Common",
+        "Base",
+        "Gutil",
+        "StarRocksGen"
+      ],
+      "allowed_test_targets": ["data_workflows_test"],
+      "allowed_test_link_deps": [
+        "DataWorkflows",
+        "Storage",
+        "Rowset",
+        "Exec",
+        "Runtime",
+        "StorageBase",
+        "StoragePrimitive",
+        "ExecRuntime",
+        "ExecPrimitive",
+        "ComputeEnv",
+        "Serde",
+        "RuntimeEnv",
+        "RuntimeCore",
+        "Platform",
+        "FSCore",
+        "IO",
+        "ChunkCore",
+        "ColumnCore",
+        "Types",
+        "Common",
+        "Base",
+        "Gutil",
+        "StarRocksGen"
+      ],
+      "remediation": "Keep DataWorkflows limited to data-plane orchestration over concrete Storage and Exec; move service/bootstrap, agent scheduling, connector registration, and lower reusable primitives to their owning layers."
     },
     {
       "id": "exprcore",
